requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
browser-use>=0.1.0
langchain>=0.1.0 